backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.7.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0   # JWT
python-multipart==0.0.9    # file uploads
httpx==0.27.0              # async HTTP (Pinata uploads)
py-algorand-sdk             # Algorand Python SDK
python-dotenv==1.0.1
